// src/app/features/card/models/card-detail.model.ts
export type CardDetail = {
  category: string;
  id: string;
  illustrator: string;
  image: string;
  localId: string;
  name: string;
  rarity: string;
  set: {
    cardCount: {
      official: number;
      total: number;
    };
    id: string;
    logo: string;
    name: string;
    symbol: string;
  };
  variants: {
    firstEdition: boolean;
    holo: boolean;
    normal: boolean;
    reverse: boolean;
    wPromo: boolean;
  };
  hp: number;
  types: string[];
  evolveFrom?: string;
  description: string;
  stage: string;
  attacks: {
    cost: string[];
    name: string;
    effect: string;
    damage?: number;
  }[];
  weaknesses: {
    type: string;
    value: string;
  }[];
  resistances: {
    type: string;
    value: string;
  }[];
  retreat: number;
  regulationMark: string;
  legal: {
    standard: boolean;
    expanded: boolean;
  };
}